requirements.txt
Django==2.2.6

psycopg2==2.8.4

django-braces==1.13.0
djangorestframework==3.10.3
django-filter==2.2.0

requests==2.22.0

six==1.12.0

boto3==1.10.41

django-cors-headers==3.1.1
django-simple-history==2.8.0

coverage==4.5.4
django-rest-swagger==2.2.0
pylint==2.4.3

elasticsearch==7.0.5
jsonschema==3.1.1

uwsgi==2.0.15

git+https://github.com/arpan29/mew

python-slugify==4.0.0
